+++ pyrps/urls.py
@@ -7,6 +7,7 @@ urlpatterns = [
     path('', include('main.urls')),
     path('dnd5/', include('dnd5.urls')),
     path('smb/', include('smb.urls')),
+    path('soundboard/', include('soundboard.urls')),
     path('admin/', admin.site.urls),
 ]