+++ smb/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
-from .models import Mech, MechClass
+from .models import Match, Mech, MechClass
 
+admin.site.register(Match)
 admin.site.register(Mech)
 admin.site.register(MechClass)