+++ smb/models.py
@@ -15,9 +15,9 @@ from . rules import ability_level
 # Globals
 default_stock = 3
 
-# CHARACTERS
+# CLASSES
 
-## Mech Class
+''' Mech Class '''
 class MechClass(models.Model):
   objid = models.CharField(max_length=12, unique=True)
   name = models.CharField(max_length=255, unique=True)
@@ -45,7 +45,7 @@ class MechClass(models.Model):
   def __str__(self):
     return self.name
 
-## Character
+''' Character '''
 class Mech(models.Model):
   ''' Character Info '''
   mech_id = models.IntegerField(default=utils.gen_charid(),
@@ -57,20 +57,20 @@ class Mech(models.Model):
     on_delete=models.SET_NULL)
 
   ''' Mech Build '''
-  strength = models.IntegerField(default=10, 
-    validators=[MinValueValidator(1), MaxValueValidator(99)])
-  defense = models.IntegerField(default=10, 
-    validators=[MinValueValidator(1), MaxValueValidator(99)])
-  dexterity = models.IntegerField(default=10, 
-    validators=[MinValueValidator(1), MaxValueValidator(99)])
-  hull = models.IntegerField(default=10, 
-    validators=[MinValueValidator(1), MaxValueValidator(99)])
-  repair_kits = models.IntegerField(default=10, 
-    validators=[MinValueValidator(1), MaxValueValidator(99)])
-  shields = models.IntegerField(default=10, 
-    validators=[MinValueValidator(1), MaxValueValidator(99)])
-  skill_points = models.IntegerField(default=10, 
-    validators=[MinValueValidator(1), MaxValueValidator(99)])
+  strength = models.IntegerField(default=0, 
+    validators=[MinValueValidator(0), MaxValueValidator(99)])
+  defense = models.IntegerField(default=0, 
+    validators=[MinValueValidator(0), MaxValueValidator(99)])
+  dexterity = models.IntegerField(default=0, 
+    validators=[MinValueValidator(0), MaxValueValidator(99)])
+  hull = models.IntegerField(default=0, 
+    validators=[MinValueValidator(0), MaxValueValidator(99)])
+  repair_kits = models.IntegerField(default=0, 
+    validators=[MinValueValidator(0), MaxValueValidator(99)])
+  shields = models.IntegerField(default=0, 
+    validators=[MinValueValidator(0), MaxValueValidator(99)])
+  skill_points = models.IntegerField(default=0, 
+    validators=[MinValueValidator(0), MaxValueValidator(99)])
 
   ''' Inventory '''
 
@@ -85,6 +85,8 @@ class Mech(models.Model):
   ''' DB Info '''
   date_added = models.DateTimeField('Date Added', auto_now_add=True)
   date_modified = models.DateTimeField('Date Modified', auto_now=True)
+
+  hull_points = 100
 
 
   ''' Mech Stats '''
@@ -104,17 +106,17 @@ class Mech(models.Model):
   def total_hull(self):
     return self.get_stat_total(self.hull, self.mech_class.base_hull)
 
-  def total_repair_kits(self):
+  def stock_repair_kits(self):
     total = self.get_stat_total(self.repair_kits, self.mech_class.base_repair_kits)
     total = total + default_stock
     return total
 
-  def total_shields(self):
+  def stock_shields(self):
     total = self.get_stat_total(self.shields, self.mech_class.base_shields)
     total = total + default_stock
     return total
 
-  def total_skill_points(self):
+  def stock_skill_points(self):
     total = self.get_stat_total(self.skill_points, self.mech_class.base_skill_points)
     total = total + default_stock
     return total
@@ -124,8 +126,6 @@ class Mech(models.Model):
     for a in ability_level:
       if score in a['score']:
         hit_die = a['die']
-        print('HIT!')
-        print(a['die'])
       elif score > 10:
         hit_die = '1d20'
       elif score < 0:
@@ -141,12 +141,45 @@ class Mech(models.Model):
   def dexterity_hit_die(self):
     return self.get_hit_die(self.total_dexterity())
 
+  def hull_range(self):
+    int_range = utils.get_range(self.total_hull())
+    return int_range
+
+  def repair_kit_range(self):
+    int_range = utils.get_range(self.stock_repair_kits())
+    return int_range
+
+  def shields_range(self):
+    int_range = utils.get_range(self.stock_shields())
+    return int_range
+
+  def skill_points_range(self):
+    int_range = utils.get_range(self.stock_skill_points())
+    return int_range
+
   def max_hit_points(self):
-    return 1000 + (self.total_hull() * 100)
+    return 1000 + (self.total_hull() * 50)
 
   ''' Meta '''
   class Meta:
     ordering = ["name"]
+
+  def __str__(self):
+    return self.name
+
+
+
+''' Match '''
+class Match(models.Model):
+  name = models.CharField(max_length=255, unique=True)
+  mechs = models.ManyToManyField(Mech)
+  date_added = models.DateTimeField('Date Added', auto_now_add=True)
+  date_modified = models.DateTimeField('Date Modified', auto_now_add=True)
+
+  ''' Meta '''
+  class Meta:
+    ordering = ["date_modified"]
+    verbose_name_plural = "Matches"
 
   def __str__(self):
     return self.name