+++ soundboard/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+from .models import Music, Sfx, Soundboard 
+
+admin.site.register(Music)
+admin.site.register(Sfx)
+admin.site.register(Soundboard)