+++ soundboard/templatetags/soundboard.py
@@ -3,12 +3,12 @@ import sys
 from django import template
 from django.template.defaultfilters import stringfilter
 from django.conf import settings
-from smb.apps import SmbConfig
+from django.utils.safestring import mark_safe
+from soundboard.apps import SoundboardConfig
 
 register = template.Library()
 
 # Engine title
 @register.simple_tag
 def engine_title():
-  # engine_title = 'Super Mech Battle'
-  return SmbConfig.verbose_name
+  return SoundboardConfig.verbose_name